Revert "moving check for snackbar to Posts"
This reverts commit f4607d22f358d4af96784fc3292ac2f4fc909789.
